@@ -1,8 +1,6 @@
 module github.com/logbn/zongzi
 
-go 1.24.0
-
-toolchain go1.24.3
+go 1.25.1
 
 require (
 	github.com/benbjohnson/clock v1.3.5
