@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             v3.12.4
 // source: zongzi.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	Internal_Ping_FullMethodName    = "/zongzi.Internal/Ping"
@@ -30,11 +30,14 @@ const (
 	Internal_Commit_FullMethodName  = "/zongzi.Internal/Commit"
 	Internal_Read_FullMethodName    = "/zongzi.Internal/Read"
 	Internal_Watch_FullMethodName   = "/zongzi.Internal/Watch"
+	Internal_Stream_FullMethodName  = "/zongzi.Internal/Stream"
 )
 
 // InternalClient is the client API for Internal service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Zongzi uses an internal gRPC service for cluster coordination
 type InternalClient interface {
 	// Ping is a noop for timing purposes
 	Ping(ctx context.Context, in *PingRequest, opts ...grpc.CallOption) (*PingResponse, error)
@@ -62,7 +65,9 @@ type InternalClient interface {
 	// Read provides unary request/response query forwarding
 	Read(ctx context.Context, in *ReadRequest, opts ...grpc.CallOption) (*ReadResponse, error)
 	// Watch provides streaming query response forwarding
-	Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (Internal_WatchClient, error)
+	Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[WatchResponse], error)
+	// Stream provides streaming proposal forwarding
+	Stream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[StreamRequest, StreamResponse], error)
 }
 
 type internalClient struct {
@@ -74,8 +79,9 @@ func NewInternalClient(cc grpc.ClientConnInterface) InternalClient {
 }
 
 func (c *internalClient) Ping(ctx context.Context, in *PingRequest, opts ...grpc.CallOption) (*PingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingResponse)
-	err := c.cc.Invoke(ctx, Internal_Ping_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Ping_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -83,8 +89,9 @@ func (c *internalClient) Ping(ctx context.Context, in *PingRequest, opts ...grpc
 }
 
 func (c *internalClient) Probe(ctx context.Context, in *ProbeRequest, opts ...grpc.CallOption) (*ProbeResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ProbeResponse)
-	err := c.cc.Invoke(ctx, Internal_Probe_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Probe_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -92,8 +99,9 @@ func (c *internalClient) Probe(ctx context.Context, in *ProbeRequest, opts ...gr
 }
 
 func (c *internalClient) Info(ctx context.Context, in *InfoRequest, opts ...grpc.CallOption) (*InfoResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(InfoResponse)
-	err := c.cc.Invoke(ctx, Internal_Info_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Info_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -101,8 +109,9 @@ func (c *internalClient) Info(ctx context.Context, in *InfoRequest, opts ...grpc
 }
 
 func (c *internalClient) Members(ctx context.Context, in *MembersRequest, opts ...grpc.CallOption) (*MembersResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MembersResponse)
-	err := c.cc.Invoke(ctx, Internal_Members_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Members_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -110,8 +119,9 @@ func (c *internalClient) Members(ctx context.Context, in *MembersRequest, opts .
 }
 
 func (c *internalClient) Join(ctx context.Context, in *JoinRequest, opts ...grpc.CallOption) (*JoinResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(JoinResponse)
-	err := c.cc.Invoke(ctx, Internal_Join_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Join_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -119,8 +129,9 @@ func (c *internalClient) Join(ctx context.Context, in *JoinRequest, opts ...grpc
 }
 
 func (c *internalClient) Add(ctx context.Context, in *AddRequest, opts ...grpc.CallOption) (*AddResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AddResponse)
-	err := c.cc.Invoke(ctx, Internal_Add_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Add_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -128,8 +139,9 @@ func (c *internalClient) Add(ctx context.Context, in *AddRequest, opts ...grpc.C
 }
 
 func (c *internalClient) Index(ctx context.Context, in *IndexRequest, opts ...grpc.CallOption) (*IndexResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(IndexResponse)
-	err := c.cc.Invoke(ctx, Internal_Index_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Index_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -137,8 +149,9 @@ func (c *internalClient) Index(ctx context.Context, in *IndexRequest, opts ...gr
 }
 
 func (c *internalClient) Apply(ctx context.Context, in *ApplyRequest, opts ...grpc.CallOption) (*ApplyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ApplyResponse)
-	err := c.cc.Invoke(ctx, Internal_Apply_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Apply_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -146,8 +159,9 @@ func (c *internalClient) Apply(ctx context.Context, in *ApplyRequest, opts ...gr
 }
 
 func (c *internalClient) Commit(ctx context.Context, in *CommitRequest, opts ...grpc.CallOption) (*CommitResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CommitResponse)
-	err := c.cc.Invoke(ctx, Internal_Commit_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Commit_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -155,20 +169,22 @@ func (c *internalClient) Commit(ctx context.Context, in *CommitRequest, opts ...
 }
 
 func (c *internalClient) Read(ctx context.Context, in *ReadRequest, opts ...grpc.CallOption) (*ReadResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ReadResponse)
-	err := c.cc.Invoke(ctx, Internal_Read_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Internal_Read_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *internalClient) Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (Internal_WatchClient, error) {
-	stream, err := c.cc.NewStream(ctx, &Internal_ServiceDesc.Streams[0], Internal_Watch_FullMethodName, opts...)
+func (c *internalClient) Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[WatchResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Internal_ServiceDesc.Streams[0], Internal_Watch_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &internalWatchClient{stream}
+	x := &grpc.GenericClientStream[WatchRequest, WatchResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -178,26 +194,27 @@ func (c *internalClient) Watch(ctx context.Context, in *WatchRequest, opts ...gr
 	return x, nil
 }
 
-type Internal_WatchClient interface {
-	Recv() (*WatchResponse, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Internal_WatchClient = grpc.ServerStreamingClient[WatchResponse]
 
-type internalWatchClient struct {
-	grpc.ClientStream
-}
-
-func (x *internalWatchClient) Recv() (*WatchResponse, error) {
-	m := new(WatchResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
+func (c *internalClient) Stream(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[StreamRequest, StreamResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &Internal_ServiceDesc.Streams[1], Internal_Stream_FullMethodName, cOpts...)
+	if err != nil {
 		return nil, err
 	}
-	return m, nil
+	x := &grpc.GenericClientStream[StreamRequest, StreamResponse]{ClientStream: stream}
+	return x, nil
 }
+
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Internal_StreamClient = grpc.BidiStreamingClient[StreamRequest, StreamResponse]
 
 // InternalServer is the server API for Internal service.
 // All implementations must embed UnimplementedInternalServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Zongzi uses an internal gRPC service for cluster coordination
 type InternalServer interface {
 	// Ping is a noop for timing purposes
 	Ping(context.Context, *PingRequest) (*PingResponse, error)
@@ -225,13 +242,18 @@ type InternalServer interface {
 	// Read provides unary request/response query forwarding
 	Read(context.Context, *ReadRequest) (*ReadResponse, error)
 	// Watch provides streaming query response forwarding
-	Watch(*WatchRequest, Internal_WatchServer) error
+	Watch(*WatchRequest, grpc.ServerStreamingServer[WatchResponse]) error
+	// Stream provides streaming proposal forwarding
+	Stream(grpc.BidiStreamingServer[StreamRequest, StreamResponse]) error
 	mustEmbedUnimplementedInternalServer()
 }
 
-// UnimplementedInternalServer must be embedded to have forward compatible implementations.
-type UnimplementedInternalServer struct {
-}
+// UnimplementedInternalServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedInternalServer struct{}
 
 func (UnimplementedInternalServer) Ping(context.Context, *PingRequest) (*PingResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Ping not implemented")
@@ -263,10 +285,14 @@ func (UnimplementedInternalServer) Commit(context.Context, *CommitRequest) (*Com
 func (UnimplementedInternalServer) Read(context.Context, *ReadRequest) (*ReadResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Read not implemented")
 }
-func (UnimplementedInternalServer) Watch(*WatchRequest, Internal_WatchServer) error {
+func (UnimplementedInternalServer) Watch(*WatchRequest, grpc.ServerStreamingServer[WatchResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method Watch not implemented")
 }
+func (UnimplementedInternalServer) Stream(grpc.BidiStreamingServer[StreamRequest, StreamResponse]) error {
+	return status.Errorf(codes.Unimplemented, "method Stream not implemented")
+}
 func (UnimplementedInternalServer) mustEmbedUnimplementedInternalServer() {}
+func (UnimplementedInternalServer) testEmbeddedByValue()                  {}
 
 // UnsafeInternalServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to InternalServer will
@@ -276,6 +302,13 @@ type UnsafeInternalServer interface {
 }
 
 func RegisterInternalServer(s grpc.ServiceRegistrar, srv InternalServer) {
+	// If the following call pancis, it indicates UnimplementedInternalServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Internal_ServiceDesc, srv)
 }
 
@@ -464,21 +497,18 @@ func _Internal_Watch_Handler(srv interface{}, stream grpc.ServerStream) error {
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(InternalServer).Watch(m, &internalWatchServer{stream})
+	return srv.(InternalServer).Watch(m, &grpc.GenericServerStream[WatchRequest, WatchResponse]{ServerStream: stream})
 }
 
-type Internal_WatchServer interface {
-	Send(*WatchResponse) error
-	grpc.ServerStream
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Internal_WatchServer = grpc.ServerStreamingServer[WatchResponse]
+
+func _Internal_Stream_Handler(srv interface{}, stream grpc.ServerStream) error {
+	return srv.(InternalServer).Stream(&grpc.GenericServerStream[StreamRequest, StreamResponse]{ServerStream: stream})
 }
 
-type internalWatchServer struct {
-	grpc.ServerStream
-}
-
-func (x *internalWatchServer) Send(m *WatchResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type Internal_StreamServer = grpc.BidiStreamingServer[StreamRequest, StreamResponse]
 
 // Internal_ServiceDesc is the grpc.ServiceDesc for Internal service.
 // It's only intended for direct use with grpc.RegisterService,
@@ -533,6 +563,12 @@ var Internal_ServiceDesc = grpc.ServiceDesc{
 			StreamName:    "Watch",
 			Handler:       _Internal_Watch_Handler,
 			ServerStreams: true,
+		},
+		{
+			StreamName:    "Stream",
+			Handler:       _Internal_Stream_Handler,
+			ServerStreams: true,
+			ClientStreams: true,
 		},
 	},
 	Metadata: "zongzi.proto",
